Updated Kexts and Bootloader
Updated kexts to latest Nightly versions, and also Updated Opencore, with latest changes on config.plist. OC is now on version 0.6.5
Also added HibernationFixup.kext to the kexts folder. Hopefully Hibernation will now work properly, however on my machine I still have the time in which it will hibernate to default.  If you want to change the way it behaves, have a look in here >> https://www.howtogeek.com/260478/how-to-choose-when-your-mac-hibernates-or-enters-standby/ <<